Backend/requirements/base.txt
# Django Core
Django==5.0.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
python-decouple==3.8.*

# Database
psycopg2-binary==2.9.*
dj-database-url==2.1.*

# Celery & Redis
celery==5.3.*
redis==5.0.*
django-redis==5.4.*

# OpenAI (legacy)
openai==1.6.*

# HTTP client for OpenRouter gateway
requests>=2.31.0

# Stripe
stripe==7.10.*

# Document Generation
python-docx==1.1.*  # Best choice for DOCX generation - industry standard
fpdf2==2.7.*        # PDF generation - pure Python, no dependencies (works in containers)

# Email
django-ses==3.5.*

# Security
django-ratelimit==4.1.*
django-csp==3.8.*

# Utilities
python-dateutil==2.8.*
pytz==2023.3

# OAuth
google-auth==2.23.*
google-auth-oauthlib==1.1.*
google-auth-httplib2==0.1.*
requests-oauthlib==1.3.*

# Static files
whitenoise==6.6.*